refactor: Minimize the scope of interests for module-specific type defs
